backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
celery==5.3.6
redis==5.0.1
paramiko==3.4.0
pydantic==2.5.2
pydantic-settings==2.1.0
